+++ packages/kit/test/prerendering/disabled/svelte.config.js
@@ -1,0 +1,27 @@
+import path from 'path';
+import adapter from '../../../../adapter-static/index.js';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		adapter: adapter(),
+
+		prerender: {
+			enabled: false
+		},
+
+		vite: {
+			build: {
+				minify: false
+			},
+			clearScreen: false,
+			server: {
+				fs: {
+					allow: [path.resolve('../../../src')]
+				}
+			}
+		}
+	}
+};
+
+export default config;